accounts_service/requirements.txt
django
djangorestframework
djoser
psycopg2-binary